+++ packages/livepad/cli.js
@@ -1,0 +1,6 @@
+#!/usr/bin/env node
+'use strict';
+
+module.exports = require('@livepad/cli/cli');
+
+if (require.main === module) module.exports.main();