+++ packages/livepad/server.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = require('@livepad/cli');
+
+if (require.main === module) require('@livepad/cli/cli').main();